+++ drivers/resdisksgcp_nfs_cg/main.go
@@ -447,7 +447,11 @@ func (t *T) Status(ctx context.Context) status.T {
 		t.StatusLog().Info("xaas status disabled")
 		return status.NotApplicable
 	}
-	if sgcphelper.NeedsCacheClear() {
+	useCache, err := sgcphelper.UseCache()
+	if err != nil {
+		t.StatusLog().Warn("%s", err)
+	}
+	if !useCache {
 		if err := t.mgr.cacheClearGetCg(); err != nil {
 			t.Log().Debugf("clear get cg cache failed: %s", err)
 			t.StatusLog().Warn("possible stale value: clear get cg cache failed")